RF: simplify + robustify zenodo.update_lookup
It now should work also disregarding explicitly specified to be null or empty
ORCID entries or values for name or affiliation which people might explicitly
white out in a specific case. Before it was solely based on the presence of the
key and I think making it more value based would be more robust.

Also dictionary comprehension for lookup generation should be easier to read IMHO,
and the code simplifications/reduction of [][] lookups also should be making it
easier to read.

Also may be it is just an anecdotal evidence and just coincided, but with this code
we got one more affiliation assigned in datalad case right after prior run
http://github.com/datalad/datalad/commit/ffc008b67c0596c16dadd7daeb58204a010ece74